pyproject.toml
[project]
name = "credo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
